+++ agent/stats/engine.go
@@ -19,12 +19,12 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/aws/amazon-ecs-agent/agent/acs/model/ecstcs"
 	"github.com/aws/amazon-ecs-agent/agent/api"
 	"github.com/aws/amazon-ecs-agent/agent/config"
 	ecsengine "github.com/aws/amazon-ecs-agent/agent/engine"
 	"github.com/aws/amazon-ecs-agent/agent/logger"
 	"github.com/aws/amazon-ecs-agent/agent/stats/resolver"
+	"github.com/aws/amazon-ecs-agent/agent/tcs/model/ecstcs"
 	"golang.org/x/net/context"
 )
 
@@ -73,19 +73,6 @@ func (resolver *DockerContainerMetadataResolver) ResolveTask(dockerID string) (*
 	}
 
 	return task, nil
-}
-
-// ResolveName resolves the container name, given container id.
-func (resolver *DockerContainerMetadataResolver) ResolveName(dockerID string) (string, error) {
-	if resolver.dockerTaskEngine == nil {
-		return "", fmt.Errorf("Docker task engine uninitialized.")
-	}
-	container, found := resolver.dockerTaskEngine.State().ContainerById(dockerID)
-	if !found {
-		return "", fmt.Errorf("Could not map docker id to container")
-	}
-
-	return container.DockerName, nil
 }
 
 // NewDockerStatsEngine creates a new instance of the DockerStatsEngine object.
@@ -192,12 +179,14 @@ func (engine *DockerStatsEngine) GetInstanceMetrics() (*ecstcs.MetricsMetadata, 
 			continue
 		}
 
-		taskMetrics = append(taskMetrics, &ecstcs.TaskMetric{
-			TaskArn:               &taskArn,
+		metricTaskArn := taskArn
+		taskMetric := &ecstcs.TaskMetric{
+			TaskArn:               &metricTaskArn,
 			TaskDefinitionFamily:  &taskDef.family,
 			TaskDefinitionVersion: &taskDef.version,
 			ContainerMetrics:      containerMetrics,
-		})
+		}
+		taskMetrics = append(taskMetrics, taskMetric)
 	}
 
 	if len(taskMetrics) == 0 {
@@ -282,7 +271,6 @@ func (engine *DockerStatsEngine) addContainer(dockerID string) {
 		return
 	}
 
-	containerName, err := engine.resolver.ResolveName(dockerID)
 	if err != nil {
 		log.Debug("Could not get name for container, ignoring", "err", err, "id", dockerID)
 		return
@@ -304,7 +292,7 @@ func (engine *DockerStatsEngine) addContainer(dockerID string) {
 	}
 
 	log.Debug("Adding container to stats watch list", "id", dockerID, "task", task.Arn)
-	container := newCronContainer(&dockerID, &containerName, engine.dockerGraphPath)
+	container := newCronContainer(&dockerID, engine.dockerGraphPath)
 	engine.tasksToContainers[task.Arn][dockerID] = container
 	engine.tasksToDefinitions[task.Arn] = &taskDefinition{family: task.Family, version: task.Version}
 	container.StartStatsCron()
@@ -393,9 +381,6 @@ func (engine *DockerStatsEngine) getContainerMetricsForTask(taskArn string) ([]*
 		}
 
 		containerMetrics = append(containerMetrics, &ecstcs.ContainerMetric{
-			Metadata: &ecstcs.ContainerMetadata{
-				Name: container.containerMetadata.Name,
-			},
 			CpuStatsSet:    cpuStatsSet,
 			MemoryStatsSet: memoryStatsSet,
 		})