+++ agent/stats/mock/engine.go
@@ -18,7 +18,7 @@ package mock_stats
 
 import (
 	gomock "code.google.com/p/gomock/gomock"
-	ecstcs "github.com/aws/amazon-ecs-agent/agent/acs/model/ecstcs"
+	ecstcs "github.com/aws/amazon-ecs-agent/agent/tcs/model/ecstcs"
 )
 
 // Mock of Engine interface