+++ agent/stats/resolver/resolver.go
@@ -20,5 +20,4 @@ import "github.com/aws/amazon-ecs-agent/agent/api"
 // ContainerMetadataResolver defines methods to resolve meta-data.
 type ContainerMetadataResolver interface {
 	ResolveTask(string) (*api.Task, error)
-	ResolveName(string) (string, error)
 }