+++ agent/stats/types.go
@@ -37,7 +37,6 @@ type UsageStats struct {
 // ContainerMetadata contains meta-data information for a container.
 type ContainerMetadata struct {
 	DockerID *string `json:"-"`
-	Name     *string `json:"name"`
 }
 
 // CronContainer abstracts methods to gather and aggregate utilization data for a container.